third_party/WebKit/LayoutTests/fast/scroll-behavior/overflow-scroll-animates.html

<!DOCTYPE html>
<style>

#container {
  width: 200px;
  height: 200px;
  overflow: scroll;
}

#content {
  width: 750px;
  height: 750px;
  background-color: blue;
}

</style>
<script src="../../resources/js-test.js"></script>
<div id="container">
  <div id="content"></div>
</div>
<div id="console"></div>
<script>

jsTestIsAsync = true;
element = document.getElementById("container");

var testConfigs = [
  {preciseDeltas: false, expectSmooth: true},
  {preciseDeltas: true, expectSmooth: false},
];
var config;
var waitingForScroll = false;
var scrollStart, scrollDuration;

function nextConfig() {
    config = testConfigs.shift();
    if (!config)
        finishJSTest();

    element.addEventListener("scroll", onElementScroll);

    testRunner.displayAsyncThen(function() {
        eventSender.mouseMoveTo(100, 100);
        eventSender.mouseScrollBy(0, -2, /* paged */ false,
            config.preciseDeltas);
        scrollStart = performance.now();
    });
}

function reset() {
    element.removeEventListener("scroll", onElementScroll);
    element.scrollTop = 0;
    waitingForScroll = false;
    scrollStart = undefined;
}

function onElementScroll() {
    if (waitingForScroll) {
        if (element.scrollTop == 80) {
            scrollDuration = performance.now() - scrollStart;
            shouldBeTrue("scrollDuration < 500");
            reset();
            nextConfig();
        }
    } else if (!scrollStart) {
        return;
    } else if (config.expectSmooth) {
        shouldNotBe("element.scrollTop", "80");
        waitingForScroll = true;
    } else {
        shouldBe("element.scrollTop", "80");
        reset();
        nextConfig();
    }
}

function runTest() {
    internals.settings.setScrollAnimatorEnabled(true);
    nextConfig();
}

description("Tests that overflow scrolls are animated, unless the wheel " +
            "event has precise scrolling deltas.");

if (window.eventSender)
    runTest();
else
    debug("FAIL: This test requires window.eventSender.");

</script>
